+++ .github/workflows/tests.yml
@@ -24,4 +24,4 @@ jobs:
       - run: |
           pip install -U pip
           pip install -e .
-          pytest tests.py -s
+          pytest -s
